+++ lib/metapede_web/live/live_components/breadcrumbs_component.ex
@@ -22,22 +22,14 @@ defmodule MetapedeWeb.LiveComponents.BreadcrumbsComponent do
     </div>
     """
   end
-end
 
-defmodule MetapedeWeb.LiveComponents.Crumb do
-  use MetapedeWeb, :live_component
+  def handle_event("reset_breadcrumbs" <> index, _, socket) do
+    updated_breadcrumbs =
+      socket.assigns.breadcrumbs
+      |> Enum.take(String.to_integer(index))
 
-  def render(assigns) do
-    ~L"""
-    <div class="crumb">
-    <span>
-    <%= live_patch @crumb["name"],
-      to: Routes.time_period_show_path(@socket, :main, @crumb["id"]),
-      phx_click: "reset_breadcrumbs" <> Integer.to_string(@index)
-    %>
-    </span>
-    </div>
-    <div class="caret">></div>
-    """
+    {:noreply, socket |> assign(breadcrumbs: updated_breadcrumbs)}
   end
+
 end
+
